- nrrdAccessors.h has been removed.  No one ever needed prototypes for all those functions.
- new source files convert.c (for all the functions used in
nrrdConvert()) and arith.c (place for all functions which will do
arithmetic on nrrds)

- nrrdConvert() now actually does that I always said it did: allow you
to apply C's type cast sementics to nrrds.  This way, when you
nrrdConvert a float nrrd to an int nrrd, the mapping on the individual
values is exactly as what you'd get by doing a C cast or assignment.
This is done by actually doing an assingment.  The two-dimensional
array of conversion functions (121 in all) is created with some
pre-processor black magic, but you shouldn't have to worry about that.

- min/max stuff.  The Nrrd struct has two fields for "min" and "max".
How these should be used, and under what circumstances their values
can be trusted, is still up for a little more clarification.  But one
simple step was to declare different "behaviors" for the setting and
getting of those variables.  There are four, in the nrrdMinMax enum:

  nrrdMinMaxCalc,            /* 1: find the actual min/max of the
				data, and use it for compuation, but
				don't set them in nrrd */
  nrrdMinMaxCalcSet,         /* 2: like nrrdMinMaxCalc, but DO record
				the min/max in the nrrd */
  nrrdMinMaxUse,             /* 3: trust and use the min/max values
				already stored in the nrrd */
  nrrdMinMaxInsteadUse,      /* 4: trust and use the min/max values
				specified by some alternate means */

nrrdMinMaxFind() and the new nrrdMinMaxDo() both use these, the
only difference being that nrrdMinMaxDo() actually has a means of
exploiting min/max behavior nrrdMinMaxInsteadUse: two more arguments.

- the "measures" in the nrrdMeasr enum have been reordered slightly,
and they're no longer called "measr"s.  Now they're just "measures".
English words are a good thing.

- nrrdQuantize: The arguments to this function have changed: instead
of giving it a min and max, you use a nrrdMinMax behavior.

- some renamings:
  nrrdRange() --> nrrdMinMaxFind()
  nrrdAddComment() --> nrrdCommentAdd()
  nrrdClearComments() --> nrrdCommentClear()
  nrrdScanComments() --> nrrdCommentScan()

- The policy of when the void* pointer in nrrd should be used, free'd,
or set, is also something that should be clarified.  For the time
being, the following is most likely true: Whenever a nrrd function
gets a nrrd to put output in (often with variable name "nout"), the
space for the output in that nrrd is insured with a call to the newly
created nrrdMaybeAlloc(), which determines if the the memory size of
the current nrrd is exactly the same as the memory size desired, and
pif so, reuses that memory.  Otherwise, it is freed and then allocated
to the desired size.  The use of nrrdMaybeAlloc has been added to all
(I hope) functions which take a nrrd for output

- map.c now has nrrdApplyGamma() which is basically gamma correction
for nrrds.  Also allows inverting the values between the min and max.

- following changes in air, replaced uses of airNand() and airNanf()
with AIR_NAN.